# swagger.yaml
definitions:
    LoginRequest:
        properties:
            email:
                example: example@mail.com
                type: string
                x-go-name: Email
            password:
                type: string
                x-go-name: Password
        required:
            - email
            - password
        type: object
        x-go-package: go_auth/requests
    SignUpRequest:
        properties:
            c_password:
                type: string
                x-go-name: C_Password
            email:
                example: example@mail.com
                type: string
                x-go-name: Email
            name:
                example: User
                type: string
                x-go-name: Name
            password:
                type: string
                x-go-name: Password
        required:
            - email
            - password
            - name
            - c_password
        type: object
        x-go-package: go_auth/requests
    UserRequest:
        properties:
            email:
                type: string
                x-go-name: Email
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: go_auth/requests
    UserResponse:
        properties:
            email:
                type: string
                x-go-name: Email
            id:
                format: uint64
                type: integer
                x-go-name: ID
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: go_auth/responses
paths:
    /login:
        post:
            operationId: idLogin
            parameters:
                - description: Login a user.
                  in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/LoginRequest'
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "422":
                    $ref: '#/responses/invalidResponse'
            summary: Logs in a user.
    /sign_up:
        post:
            operationId: idSignUp
            parameters:
                - description: Signs up a new user.
                  in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/SignUpRequest'
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "422":
                    $ref: '#/responses/invalidResponse'
            summary: Signs up a new user.
    /user:
        delete:
            operationId: idDeleteUser
            parameters:
                - description: JWT Token
                  in: header
                  name: Authorization
                  type: string
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            summary: Delete user.
            tags:
                - Auth
        get:
            operationId: idGetUser
            parameters:
                - description: JWT Token
                  in: header
                  name: Authorization
                  type: string
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "400":
                    $ref: '#/responses/invalidResponse'
                "401":
                    $ref: '#/responses/invalidResponse'
            summary: Get user.
            tags:
                - Auth
        put:
            operationId: idUpdateUser
            parameters:
                - description: JWT Token
                  in: header
                  name: Authorization
                  type: string
                - description: update user.
                  in: body
                  name: Body
                  schema:
                    $ref: '#/definitions/UserRequest'
            responses:
                "200":
                    $ref: '#/responses/successResponse'
                "422":
                    $ref: '#/responses/invalidResponse'
            summary: Update user details.
            tags:
                - Auth
responses:
    invalidResponse:
        description: Invalid response.
        schema:
            properties:
                Err:
                    type: string
            type: object
    successResponse:
        description: Successful response.
        headers:
            Authorization:
                description: Authorization header
                type: string
        schema:
            $ref: '#/definitions/UserResponse'
swagger: "2.0"